Move fn replace_all to separate source file
